remove eager tagging & thumbnail regen of tags
wagtail unfortunately handles tagging in a really silly and inefficient
way, which means that doing this on save will cause page saves to take
ages.

the _right_ way to handle this would be via an async worker, but for now
we should just batch it once a day or so.