fix: don't display untranslated logbooks on the index page (#273)
When the Portugese locale was created, aliases of the English
pages were created and added to the Portugese site. This meant
that English content was displayed in the Portugese site, in
particular on the logbooks page. These untranslated aliases
are now filtered out so the logbooks page only displays
Portugese content.