feat: auto-sync translated pages on save
This does not happen automatically by default. This is a problem
as it means it's very cumbersome to make sure synchronised
fields (e.g. tags) are updated across translations.

It is made automatic by adding an after_publish_page hook that
updates the TranslationSource with the new synchronised field
values, then updates each translated version of the page.

This commit also makes sure tags display their localised names
in a few places.