Translate Home Page (#275)
* feat: translate tags and homepage

* fix: create new tags in correct locale from tag widget on page edit form

* feat: localize tag cloud building and fix frontend rendering

* feat: fix seo metadata on map page

* fix: correct tag cloud indexes if a tag does not exist

* feat: add loading spinner to map page

* fix: fixes filtering by tag when tags exist in multiple locales with the same slug

* fix: add map page seo fields migration

* fix: localize tag in page metadata sidebar

* feat: auto-sync translated pages on save

This does not happen automatically by default. This is a problem
as it means it's very cumbersome to make sure synchronised
fields (e.g. tags) are updated across translations.

It is made automatic by adding an after_publish_page hook that
updates the TranslationSource with the new synchronised field
values, then updates each translated version of the page.

This commit also makes sure tags display their localised names
in a few places.

* feat: show all content on index pages (not just localized)

---------

Co-authored-by: Joaquim de Souza <joaquim.dsouza@scalefast.com>